path

+++ encoding/nn/customize.py
@@ -97,7 +97,7 @@ class SegmentationLosses2(CrossEntropyLoss):
             not_ignore = (target!=self.ignore_index).float().unsqueeze(1)
             target_cp = target.clone()
             target_cp[target_cp==self.ignore_index] = 0
-            onehot_label = F.one_hot(target_cp, num_classes =self.nclass)
+            onehot_label = F.one_hot(target_cp, num_classes =self.nclass).float()
             loss2 = torch.sum(self.bce(pred2, onehot_label.permute(0,3,1,2))*not_ignore)/torch.sum(not_ignore)
             loss3 = super(SegmentationLosses2, self).forward(pred3, target)
             return loss1 + loss2 + self.aux_weight * loss3
